# ai-service/ai-service/openapi/openapi.json
{"openapi":"3.0.2","info":{"title":"MUED LMS AI Service","description":"AI Service for Music Education LMS","version":"0.1.0"},"paths":{"/api/v1/courses/generate":{"post":{"summary":"Generate Course","description":"コース生成エンドポイント\n\n指定されたトピックと条件に基づいて新しいコースを生成します。\n現在はモックデータを返します。","operationId":"generate_course_api_v1_courses_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseGenerationRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseGenerationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/exercise/logs":{"post":{"summary":"Create Exercise Log","description":"練習記録保存エンドポイント\n\nユーザーの練習記録を保存します。\n現在はモックデータを返します。","operationId":"create_exercise_log_api_v1_exercise_logs_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExerciseLogCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExerciseLog"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/webhooks/general":{"post":{"summary":"General Webhook Handler","description":"汎用Webhookエンドポイント\n\n様々なサービスからのイベント通知を受け取ります。","operationId":"general_webhook_handler_api_v1_webhooks_general_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WebhookEvent"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/webhooks/stripe":{"post":{"summary":"Stripe Webhook Handler","description":"Stripe Webhookエンドポイント\n\nStripeからの支払い関連イベント通知を受け取ります。","operationId":"stripe_webhook_handler_api_v1_webhooks_stripe_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/chat/messages":{"get":{"summary":"Get Chat Messages","description":"チャットメッセージ取得エンドポイント\n\n指定されたルームのチャットメッセージを取得します。\n現在はモックデータを返します。","operationId":"get_chat_messages_api_v1_chat_messages_get","parameters":[{"required":true,"schema":{"title":"Room Id","type":"string"},"name":"room_id","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":20},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatMessageList"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Create Chat Message","description":"チャットメッセージ作成エンドポイント\n\n新しいチャットメッセージを作成します。\n現在はモックレスポンスを返します。","operationId":"create_chat_message_api_v1_chat_messages_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatMessageCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatMessage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/musicxml/convert":{"post":{"summary":"Convert Musicxml","description":"MusicXML変換エンドポイント\n\nMusicXMLデータをJSON形式またはプレビュー画像に変換します。\n現在はモックデータを返します。","operationId":"convert_musicxml_api_v1_musicxml_convert_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MusicXMLConvertRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MusicXMLConvertResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","description":"ルートエンドポイント","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ChatMessage":{"title":"ChatMessage","required":["id","room_id","user_id","username","message","created_at"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"room_id":{"title":"Room Id","type":"string"},"user_id":{"title":"User Id","type":"string"},"username":{"title":"Username","type":"string"},"message":{"title":"Message","type":"string"},"created_at":{"title":"Created At","type":"string","format":"date-time"}},"description":"チャットメッセージレスポンス","example":{"id":"msg-123456","room_id":"room-abc123","user_id":"user-456","username":"田中先生","message":"こんにちは、レッスンの準備はできていますか？","created_at":"2024-05-01T12:30:45.123456"}},"ChatMessageCreate":{"title":"ChatMessageCreate","required":["room_id","user_id","username","message"],"type":"object","properties":{"room_id":{"title":"Room Id","type":"string"},"user_id":{"title":"User Id","type":"string"},"username":{"title":"Username","type":"string"},"message":{"title":"Message","type":"string"}},"description":"チャットメッセージ作成リクエスト"},"ChatMessageList":{"title":"ChatMessageList","required":["messages","total"],"type":"object","properties":{"messages":{"title":"Messages","type":"array","items":{"$ref":"#/components/schemas/ChatMessage"}},"total":{"title":"Total","type":"integer"}},"description":"チャットメッセージリストレスポンス"},"CourseGenerationRequest":{"title":"CourseGenerationRequest","required":["topic","level"],"type":"object","properties":{"topic":{"title":"Topic","type":"string","description":"コースのトピック"},"level":{"title":"Level","type":"string","description":"難易度レベル (beginner, intermediate, advanced)"},"goal":{"title":"Goal","type":"string","description":"学習目標"},"keywords":{"title":"Keywords","type":"array","items":{"type":"string"},"description":"含めるべきキーワード"}},"description":"コース生成リクエストモデル","example":{"topic":"ジャズピアノ入門","level":"beginner","goal":"基本的なジャズピアノコードと即興演奏の基礎を学ぶ","keywords":["ジャズハーモニー","コードプログレッション","即興演奏"]}},"CourseGenerationResponse":{"title":"CourseGenerationResponse","required":["course_id","title","description","level","estimated_duration_hours","modules"],"type":"object","properties":{"course_id":{"title":"Course Id","type":"string"},"title":{"title":"Title","type":"string"},"description":{"title":"Description","type":"string"},"level":{"title":"Level","type":"string"},"estimated_duration_hours":{"title":"Estimated Duration Hours","type":"number"},"modules":{"title":"Modules","type":"array","items":{"$ref":"#/components/schemas/CourseModule"}}},"description":"コース生成レスポンスモデル","example":{"course_id":"jazz-piano-basics-001","title":"ジャズピアノ入門: 基礎から即興演奏まで","description":"ジャズピアノの基本的な技術とハーモニーを学び、簡単な即興演奏ができるようになるためのコースです。","level":"beginner","estimated_duration_hours":12.5,"modules":[{"title":"ジャズハーモニーの基礎","description":"ジャズで使われる基本的なコードとスケールを学びます。","order":1,"lessons":[{"title":"ジャズピアノ入門","description":"ジャズピアノの基本とコースの概要","content_type":"video","duration_minutes":15,"content_url":"https://example.com/videos/jazz-intro"}]}]}},"CourseModule":{"title":"CourseModule","required":["title","description","order","lessons"],"type":"object","properties":{"title":{"title":"Title","type":"string"},"description":{"title":"Description","type":"string"},"order":{"title":"Order","type":"integer"},"lessons":{"title":"Lessons","type":"array","items":{"$ref":"#/components/schemas/LessonContent"}}},"description":"コースモジュールモデル"},"ExerciseLog":{"title":"ExerciseLog","required":["id","user_id","instrument","duration_minutes","difficulty","date","created_at"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"user_id":{"title":"User Id","type":"string"},"instrument":{"title":"Instrument","type":"string"},"duration_minutes":{"title":"Duration Minutes","type":"integer"},"difficulty":{"title":"Difficulty","type":"string"},"notes":{"title":"Notes","type":"string"},"mood":{"title":"Mood","type":"string"},"date":{"title":"Date","type":"string","format":"date-time"},"created_at":{"title":"Created At","type":"string","format":"date-time"}},"description":"練習記録レスポンス"},"ExerciseLogCreate":{"title":"ExerciseLogCreate","required":["user_id","instrument","duration_minutes","difficulty"],"type":"object","properties":{"user_id":{"title":"User Id","type":"string"},"instrument":{"title":"Instrument","type":"string"},"duration_minutes":{"title":"Duration Minutes","type":"integer"},"difficulty":{"title":"Difficulty","type":"string"},"notes":{"title":"Notes","type":"string"},"mood":{"title":"Mood","type":"string"},"date":{"title":"Date","type":"string","format":"date-time"}},"description":"練習記録作成リクエスト"},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"LessonContent":{"title":"LessonContent","required":["title","description","content_type","duration_minutes"],"type":"object","properties":{"title":{"title":"Title","type":"string"},"description":{"title":"Description","type":"string"},"content_type":{"title":"Content Type","type":"string","description":"コンテンツタイプ (video, text, exercise)"},"duration_minutes":{"title":"Duration Minutes","type":"integer"},"content_url":{"title":"Content Url","type":"string"},"content_text":{"title":"Content Text","type":"string"}},"description":"レッスンコンテンツモデル"},"MusicXMLConvertRequest":{"title":"MusicXMLConvertRequest","required":["xml_content","format"],"type":"object","properties":{"xml_content":{"title":"Xml Content","type":"string","description":"MusicXMLコンテンツ（Base64エンコード）"},"format":{"title":"Format","type":"string","description":"変換フォーマット (json, preview)"},"options":{"title":"Options","type":"object","description":"変換オプション"}},"description":"MusicXML変換リクエスト","example":{"xml_content":"PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0iVVRGLTgiPz4KPCEtLSBNdXNpY1hNTCBTYW1wbGUgLS0+CjxzY29yZS1wYXJ0aXdpc2U+PC9zY29yZS1wYXJ0aXdpc2U+","format":"json","options":{"includeMetadata":true,"includeNotations":true}}},"MusicXMLConvertResponse":{"title":"MusicXMLConvertResponse","required":["id","format","result","created_at"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"format":{"title":"Format","type":"string"},"result":{"title":"Result","type":"object"},"preview_url":{"title":"Preview Url","type":"string"},"created_at":{"title":"Created At","type":"string","format":"date-time"}},"description":"MusicXML変換レスポンス","example":{"id":"convert-123456","format":"json","result":{"metadata":{"title":"楽譜タイトル","composer":"作曲者名"},"parts":[{"id":"P1","name":"Piano","measures":[]}]},"preview_url":"https://example.com/preview/123456.png","created_at":"2024-05-01T12:30:45.123456"}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}},"WebhookEvent":{"title":"WebhookEvent","required":["id","type","created","data"],"type":"object","properties":{"id":{"title":"Id","type":"string","description":"イベントID"},"type":{"title":"Type","type":"string","description":"イベントタイプ"},"created":{"title":"Created","type":"string","description":"作成日時","format":"date-time"},"data":{"title":"Data","type":"object","description":"イベントデータ"}},"description":"汎用Webhookイベントモデル"}}}}