# ai-service/app/api.py
"""
MUED LMS AI Service - API Endpoints
"""
from fastapi import APIRouter, HTTPException, Request
import uuid
from datetime import datetime
import json
import logging
import base64

from app.models import (
    CourseGenerationRequest, 
    CourseGenerationResponse, 
    CourseModule, 
    LessonContent, 
    ExerciseLogCreate, 
    ExerciseLog,
    WebhookEvent,
    StripeWebhookEvent,
    ChatMessageCreate,
    ChatMessage,
    ChatMessageList,
    MusicXMLConvertRequest,
    MusicXMLConvertResponse
)

router = APIRouter()

# ロガーの設定
logger = logging.getLogger("mued.api")

@router.post("/courses/generate", response_model=CourseGenerationResponse)
async def generate_course(request: CourseGenerationRequest) -> CourseGenerationResponse:
    """
    コース生成エンドポイント
    
    指定されたトピックと条件に基づいて新しいコースを生成します。
    現在はモックデータを返します。
    """
    # TODO: 実際のAIモデルを使用したコース生成を実装
    # 現段階ではモックレスポンスを返す
    
    # リクエストデータに基づいて動的にモックデータを生成
    course_id = f"{'-'.join(request.topic.lower().split()[:3])}-{uuid.uuid4().hex[:8]}"
    
    # モジュール1
    module1_lessons = [
        LessonContent(
            title=f"{request.topic}の基本概念",
            description=f"{request.topic}における基本的な概念と用語を学びます。",
            content_type="video",
            duration_minutes=15,
            content_url=f"https://example.com/videos/{course_id}/intro"
        ),
        LessonContent(
            title="歴史と背景",
            description=f"{request.topic}の歴史的背景と発展について学びます。",
            content_type="text",
            duration_minutes=20,
            content_text="ここにテキストコンテンツが入ります。"
        )
    ]
    
    # モジュール2
    module2_lessons = [
        LessonContent(
            title=f"{request.topic}の実践テクニック",
            description="実践的なテクニックと演習",
            content_type="video",
            duration_minutes=25,
            content_url=f"https://example.com/videos/{course_id}/techniques"
        ),
        LessonContent(
            title="演習問題",
            description="学んだ内容を確認するための演習問題",
            content_type="exercise",
            duration_minutes=30,
            content_url=f"https://example.com/exercises/{course_id}/practice"
        )
    ]
    
    # レスポンスの作成
    response = CourseGenerationResponse(
        course_id=course_id,
        title=f"{request.topic}: {request.level}レベルコース",
        description=f"{request.topic}の{request.level}レベルの学習者向けコースです。" + 
                   (f"目標: {request.goal}" if request.goal else ""),
        level=request.level,
        estimated_duration_hours=2.5,
        modules=[
            CourseModule(
                title="基礎と概念",
                description=f"{request.topic}の基礎知識を学びます。",
                order=1,
                lessons=module1_lessons
            ),
            CourseModule(
                title="実践とスキル習得",
                description="実践的なスキルを身につけます。",
                order=2,
                lessons=module2_lessons
            )
        ]
    )
    
    return response

@router.post("/exercise/logs", response_model=ExerciseLog)
async def create_exercise_log(request: ExerciseLogCreate) -> ExerciseLog:
    """
    練習記録保存エンドポイント
    
    ユーザーの練習記録を保存します。
    現在はモックデータを返します。
    """
    # 実際にはデータベースに保存する処理を実装
    # 現段階ではモックレスポンスを返す
    
    # 日付が設定されていない場合は現在時刻を使用
    log_date = request.date or datetime.now()
    
    # レスポンスの作成
    response = ExerciseLog(
        id=str(uuid.uuid4()),
        user_id=request.user_id,
        instrument=request.instrument,
        duration_minutes=request.duration_minutes,
        difficulty=request.difficulty,
        notes=request.notes,
        mood=request.mood,
        date=log_date,
        created_at=datetime.now()
    )
    
    return response

@router.post("/webhooks/general")
async def general_webhook_handler(event: WebhookEvent):
    """
    汎用Webhookエンドポイント
    
    様々なサービスからのイベント通知を受け取ります。
    """
    logger.info(f"Received webhook event: {event.type}")
    
    # イベントタイプに応じた処理を実装
    # 現段階では受信したイベントをログに記録するのみ
    
    return {"status": "success", "message": f"Event {event.id} of type {event.type} processed"}

@router.post("/webhooks/stripe")
async def stripe_webhook_handler(request: Request):
    """
    Stripe Webhookエンドポイント
    
    Stripeからの支払い関連イベント通知を受け取ります。
    """
    # リクエストボディを取得
    body = await request.body()
    payload = json.loads(body)
    
    # Stripe署名検証はここでは省略（本番実装時に追加）
    
    event_type = payload.get("type", "unknown")
    event_id = payload.get("id", "unknown")
    
    logger.info(f"Received Stripe webhook: {event_type} (ID: {event_id})")
    
    # イベントタイプに応じた処理
    if event_type == "payment_intent.succeeded":
        # 支払い成功時の処理
        logger.info("Payment succeeded!")
    elif event_type == "checkout.session.completed":
        # チェックアウト完了時の処理
        logger.info("Checkout completed!")
    elif event_type == "customer.subscription.created":
        # サブスクリプション作成時の処理
        logger.info("Subscription created!")
    else:
        logger.info(f"Unhandled event type: {event_type}")
    
    return {"status": "success", "message": f"Stripe event {event_id} processed"}

@router.get("/chat/messages", response_model=ChatMessageList)
async def get_chat_messages(room_id: str, limit: int = 20) -> ChatMessageList:
    """
    チャットメッセージ取得エンドポイント
    
    指定されたルームのチャットメッセージを取得します。
    現在はモックデータを返します。
    """
    # 現段階ではモックデータを返す
    messages = [
        ChatMessage(
            id=str(uuid.uuid4()),
            room_id=room_id,
            user_id="user1",
            username="田中先生",
            message="こんにちは、レッスンの準備はできていますか？",
            created_at=datetime.now()
        ),
        ChatMessage(
            id=str(uuid.uuid4()),
            room_id=room_id,
            user_id="user2",
            username="山田さん",
            message="はい、準備できています。今日は何から始めますか？",
            created_at=datetime.now()
        )
    ]
    
    return ChatMessageList(messages=messages, total=len(messages))

@router.post("/chat/messages", response_model=ChatMessage)
async def create_chat_message(message: ChatMessageCreate) -> ChatMessage:
    """
    チャットメッセージ作成エンドポイント
    
    新しいチャットメッセージを作成します。
    現在はモックレスポンスを返します。
    """
    # メッセージにIDと作成日時を追加
    message_id = str(uuid.uuid4())
    created_at = datetime.now()
    
    saved_message = ChatMessage(
        id=message_id,
        room_id=message.room_id,
        user_id=message.user_id,
        username=message.username,
        message=message.message,
        created_at=created_at
    )
    
    return saved_message

@router.post("/musicxml/convert", response_model=MusicXMLConvertResponse)
async def convert_musicxml(request: MusicXMLConvertRequest) -> MusicXMLConvertResponse:
    """
    MusicXML変換エンドポイント
    
    MusicXMLデータをJSON形式またはプレビュー画像に変換します。
    現在はモックデータを返します。
    """
    logger.info(f"MusicXML conversion requested in format: {request.format}")
    
    # 最初にフォーマットの検証を行い、無効な場合は早期に400エラーを返す
    if request.format not in ["json", "preview"]:
        logger.warning(f"Invalid format requested: {request.format}")
        raise HTTPException(status_code=400, detail=f"Unsupported format: {request.format}")
    
    try:
        # Base64デコードする（実際の処理では使用）
        # xml_content = base64.b64decode(request.xml_content).decode('utf-8')
        # ここでXML解析と変換処理を行う（現段階ではスキップ）
        
        result = {}
        preview_url = None
        
        if request.format == "json":
            # JSONフォーマットの場合
            result = {
                "metadata": {
                    "title": "サンプル楽譜",
                    "composer": "MUED LMS",
                },
                "parts": [
                    {
                        "id": "P1",
                        "name": "Piano",
                        "measures": [
                            {
                                "number": 1,
                                "notes": [
                                    {"pitch": "C4", "duration": "quarter"},
                                    {"pitch": "D4", "duration": "quarter"},
                                    {"pitch": "E4", "duration": "quarter"},
                                    {"pitch": "F4", "duration": "quarter"}
                                ]
                            }
                        ]
                    }
                ]
            }
        elif request.format == "preview":
            # プレビュー画像の場合
            result = {"status": "generated"}
            preview_url = f"https://example.com/musicxml/preview/{uuid.uuid4().hex}.png"
        
        response = MusicXMLConvertResponse(
            id=f"convert-{uuid.uuid4().hex[:8]}",
            format=request.format,
            result=result,
            preview_url=preview_url,
            created_at=datetime.now()
        )
        
        return response
        
    except Exception as e:
        logger.error(f"Error converting MusicXML: {str(e)}")
        raise HTTPException(status_code=500, detail=f"Conversion error: {str(e)}") 